Implemented miniature version of gpsd.
gpsd wasn't working with the WiFi dongle we're using for the rPi, so I wrote code to mimic its effects, called usbgps.c. It uses UNIX sockets for IPC; usbgps reads the data from the terminal, and sends it to whatever client connects to the right socket "serial_rpigps_data".